requirements.txt
streamlit
pandas
sqlalchemy
pdfplumber
python-docx
transformers
huggingface_hub
cryptography
tf-keras
fpdf
